[test run]

+++ ft_gcd.c
@@ -1,4 +1,4 @@
-int ft_gcd(int a, int b) {
+long long ft_gcd(long long a, long long b) {
     while (b != 0){
         a = b;
         b = a % b;
